// old_WeatherAlerts/ClientApp/src/app/weather-api/interfaces/ialert-properties.ts
export interface IAlertProperties {
    id: string;
    status: string;
    category: string;
    severity: string;
    urgency: string;
    event: string;
    senderName: string;
    headline: string;
    description: string;
    response: string;
    messageType: string;
    certainty: string;
    areaDesc: string;
    sent: Date;
    effective: Date;
    onset: Date;
    expires: Date;
    ends: Date;
    areas?: string[];
    zoneId?: string;
}